ROOTFS: Add Linux perf building
Since perf is a part of Linux kernel to build perf we need first
to enable building of Linux kernel.

In its turn most if not all ARC defconfigs expect to have external
folder with initramfs. Which means we have to set
BR2_TARGET_ROOTFS_INITRAMFS which silently overrides
BR2_TARGET_ROOTFS_CPIO which in its turn we need to install "init"
script which in its turn does very simple mount of devtmpfs to /dev and
sets up console (which doesn't happen automatically in case of initramfs).

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>